Refactor case attributes discovery package
