+++ src/pix_framework/discovery/case_attribute/discovery.py
