+++ tests/pix_framework/discovery/case_attribute/test_case_attribute_discovery.py
@@ -1,5 +1,5 @@
 import pandas as pd
-from pix_framework.discovery.case_attribute_discovery.discovery import discover_case_attributes
+from pix_framework.discovery.case_attribute.discovery import discover_case_attributes
 from pix_framework.io.event_log import DEFAULT_CSV_IDS
